pyproject.toml
[project]
name = "notsolocalfinance"
version = "0.1.0"
description = "Personal Finance Assistant with AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.1",
    "pandas==2.2.2",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.7.0",
    "python-multipart==0.0.9",
    "pdfplumber==0.11.2",
    "qdrant-client==1.10.1",
    "pydantic-ai>=0.0.18",
    "logfire",
    "python-dotenv",
    "fastembed",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
